chore: move `compute_merkle_root` back into `pwg/merkle.rs`
